fix: refactoring하였다.
1. domain패키지와 view패키지를 분리했다.
