+++ src/main/kotlin/racingcar/StartRaing.kt
@@ -1,5 +1,9 @@
 package racingcar
 
+import racingcar.domain.Game
+import racingcar.view.InputView
+import racingcar.view.ResultView
+
 fun main(args: Array<String>) {
     try {
         startRace()
