+++ src/main/kotlin/racingcar/domain/Game.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.domain
 
 class Game(carNames: List<String>, moveTime: Int) {
     var cars = makeCars(carNames)
