+++ src/main/kotlin/racingcar/view/InputView.kt
@@ -1,4 +1,4 @@
-package racingcar
+package racingcar.view
 
 object InputView {
     fun readCarName(): List<String> {
